Finish semantic metric view migration
