+++ internal/query/types.go
@@ -32,6 +32,30 @@ type Request struct {
 	Limit      int
 }
 
+type RowRequest struct {
+	MetricView string
+	Dimensions []Field
+	Measures   []Field
+	Filters    []Filter
+	Sort       []Sort
+	Limit      int
+	Offset     int
+}
+
+type RawValueRequest struct {
+	MetricView string
+	Dimensions []Field
+	Measure    Field
+	Filters    []Filter
+	Sort       []Sort
+	Limit      int
+}
+
+type CountRequest struct {
+	MetricView string
+	Filters    []Filter
+}
+
 type Plan struct {
 	SQL     string
 	Args    []any
